// generated/src/app/api-models/agent-license-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AgentLicenseDetail {
  seqAgentLicenseDetail: number;
  seqAgentLicenseHeader: number;
  seqAgentId: number;
  lineOfBusiness: string;
  effectiveDate: Date;
  termDate: Date;
  termReason: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}